# docs/raw/openapi.yaml
openapi: 3.1.0
info:
  title: 부여ON API
  version: 1.0.0
  description: |
    관광데이터 유즈케이스와 도메인 모델 및 규칙을 기준으로 작성한 API 명세다.
    모든 응답은 success와 data를 가지는 공통 객체 형식을 사용한다.

    멱등성 헤더가 필요한 요청은 같은 회원, 같은 작업, 같은 Idempotency-Key, 같은 요청 본문으로
    다시 호출하면 최초 성공 응답과 같은 상태 코드 및 응답 본문을 반환한다.
    같은 키를 다른 요청 본문이나 작업에 재사용하면 409 IDEMPOTENCY_KEY_REUSED를 반환한다.
    이미 완료된 행동을 다른 키로 다시 요청해 허용되지 않은 상태 전이가 발생하면 상태별 오류 코드와 함께 409를 반환한다.
servers:
  - url: https://api.buyeoon.example.com/v1
    description: Production placeholder
...
security:
  - bearerAuth: []
paths:
...
  /citizen-cards/options:
    get:
      tags: [CitizenCards]
      summary: 군민증 캐릭터와 테마 목록 조회
      description: 승인된 캐릭터와 테마를 각 카탈로그의 표시 순서대로 반환한다.
      operationId: getCitizenCardOptions
      responses:
        '200':
          $ref: '#/components/responses/CitizenCardOptionsSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /citizen-cards/location-verification:
    post:
      tags: [CitizenCards]
      summary: 군민증 발급 전 위치 확인
      description: 군민증 발급 전에 현재 위치가 부여 행정구역 안인지 확인한다. 군민증 생성 시에도 위치를 다시 검증한다.
      operationId: verifyCitizenCardLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        '200':
          $ref: '#/components/responses/CitizenCardLocationVerificationSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CitizenCardCreationForbidden'
  /citizen-cards:
    post:
      tags: [CitizenCards]
      summary: 디지털 군민증 생성
      description: 신규 회원이 필수 약관 동의를 완료하고 현재 위치가 부여 안임을 확인한 후 디지털 군민증을 생성한다.
      operationId: createCitizenCard
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CitizenCardCreateRequest'
      responses:
        '201':
          $ref: '#/components/responses/CitizenCardSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/CitizenCardCreationForbidden'
        '409':
          $ref: '#/components/responses/CitizenCardCreationConflict'
  /citizen-cards/me:
    get:
      tags: [CitizenCards]
      summary: 내 디지털 군민증 조회
      description: 현재 프로필의 표시 이름·캐릭터와 발급 시 선택한 카드 테마를 반환한다.
      operationId: getMyCitizenCard
      responses:
        '200':
          $ref: '#/components/responses/CitizenCardSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/CitizenCardNotFound'
  /citizen-cards/me/barcode:
    get:
      tags: [CitizenCards]
      summary: 군민증 바코드 조회
      description: MVP 시연용 바코드와 조회 시점의 포인트 원장 합계를 반환하며 포인트를 변경하지 않는다.
      operationId: getMyCitizenCardBarcode
      responses:
        '200':
          $ref: '#/components/responses/BarcodeSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/CitizenCardNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT의 sub는 회원 ID, sid는 인증 세션 ID다. 서버는 인증 요청마다 세션의 만료·폐기 여부를 확인한다.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 최초 성공 후 24시간 안에 같은 작업에 같은 키와 같은 요청 본문을 사용하면 최초 성공 응답을 재사용한다. 보관 중인 같은 키를 다른 작업이나 요청 본문에 재사용하면 409를 반환한다.
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
  responses:
...
    CitizenCardOptionsSuccess:
      description: 군민증 선택 항목
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CitizenCardOptionsEnvelope'
    CitizenCardLocationVerificationSuccess:
      description: 부여 내부 위치 확인 결과
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CitizenCardLocationVerificationEnvelope'
    CitizenCardSuccess:
      description: 디지털 군민증
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CitizenCardEnvelope'
    CitizenCardCreationForbidden:
      description: 현재 필수 약관 미동의 또는 부여 행정구역 외부
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            requiredTermsNotAgreed:
              summary: 현재 필수 약관 미동의
              value:
                success: false
                data:
                  code: REQUIRED_TERMS_NOT_AGREED
                  message: 현재 필수 약관에 먼저 동의해 주세요.
            outsideBuyeo:
              summary: 부여 행정구역 외부
              value:
                success: false
                data:
                  code: OUTSIDE_BUYEO
                  message: 부여 안에서만 군민증을 만들 수 있습니다.
    CitizenCardCreationConflict:
      description: 멱등성 키 재사용 또는 이미 군민증을 발급한 상태
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            idempotencyKeyReused:
              summary: 멱등성 키를 다른 요청에 재사용
              value:
                success: false
                data:
                  code: IDEMPOTENCY_KEY_REUSED
                  message: 같은 멱등성 키를 다른 요청에 재사용할 수 없습니다.
            invalidStateTransition:
              summary: 이미 군민증을 발급함
              value:
                success: false
                data:
                  code: INVALID_STATE_TRANSITION
                  message: 현재 상태에서는 요청한 작업을 수행할 수 없습니다.
    CitizenCardNotFound:
      description: 요청 회원에게 발급된 군민증이 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data:
              code: RESOURCE_NOT_FOUND
              message: 대상을 찾을 수 없습니다.
...
    BarcodeSuccess:
      description: 군민증 바코드
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BarcodeEnvelope'
...
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: 인증 필요 또는 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
    ErrorEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: false
        data:
          $ref: '#/components/schemas/ErrorData'
    ErrorData:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Location:
      type: object
      additionalProperties: false
      description: MVP에서는 유효 범위의 좌표를 요청 시점의 현재 위치로 간주하며 capturedAt은 요청 기록용이다.
      required: [latitude, longitude, capturedAt]
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        accuracyMeters:
          type: [number, 'null']
          minimum: 0
        capturedAt:
          type: string
          format: date-time
          example: '2026-08-06T15:30:00+09:00'
...
    CardOption:
      type: object
      required: [id, name, imageUrl]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        imageUrl:
          type: string
          format: uri
          description: 10분 유효한 S3 Presigned GET URL. 저장하거나 로그에 기록하지 않는다.
    CitizenCardCreateRequest:
      type: object
      additionalProperties: false
      required: [displayName, characterId, themeId, location]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 8
          description: 앞뒤 공백 제거 후 Unicode 문자 기준 1~8자이며 공백만 있거나 제어문자를 포함할 수 없다.
        characterId:
          type: string
          format: uuid
        themeId:
          type: string
          format: uuid
        location:
          $ref: '#/components/schemas/Location'
    CitizenCard:
      type: object
      required: [cardId, displayName, character, theme, issuedAt]
      properties:
        cardId:
          type: string
          format: uuid
        displayName:
          type: string
          maxLength: 8
        character:
          $ref: '#/components/schemas/CardOption'
        theme:
          $ref: '#/components/schemas/CardOption'
        issuedAt:
          type: string
          format: date-time
    BarcodeData:
      type: object
      required: [citizenCard, barcodeValue, pointBalance, simulationOnly]
      properties:
        citizenCard:
          $ref: '#/components/schemas/CitizenCard'
        barcodeValue:
          type: string
          format: uuid
          description: 서버가 생성한 UUID 형식의 MVP 시연용 값이며 결제나 포인트 차감에 사용하지 않는다.
        pointBalance:
          type: integer
          minimum: 0
        simulationOnly:
          type: boolean
          const: true
        notice:
          type: string
          example: 실제 상점에서의 사용은 제한됩니다.
...
    CitizenCardOptionsData:
      type: object
      required: [characters, themes]
      properties:
        characters:
          type: array
          description: 캐릭터 카탈로그의 표시 순서 오름차순
          items:
            $ref: '#/components/schemas/CardOption'
        themes:
          type: array
          description: 카드 테마 카탈로그의 표시 순서 오름차순
          items:
            $ref: '#/components/schemas/CardOption'
    CitizenCardOptionsEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CitizenCardOptionsData'
    CitizenCardLocationVerificationData:
      type: object
      required: [withinBuyeo]
      properties:
        withinBuyeo:
          type: boolean
          example: true
    CitizenCardLocationVerificationEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CitizenCardLocationVerificationData'
    CitizenCardEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CitizenCard'
    BarcodeEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/BarcodeData'
...